Feat/start liga

+++ src/main/java/com/adondeband/back_end_adonde_band/dominio/jornada/JornadaPort.java
@@ -5,5 +5,5 @@ import java.util.List;
 public interface JornadaPort {
     Jornada save(Jornada jornada);
 
-    List<Jornada> findById(int s);
+    List<Jornada> findById(Long s);
 }
